--- CarRental.Api/CarRental.Api.http
@@ -1,6 +1,0 @@
-@CarRental.Api_HostAddress = http://localhost:5281
-
-GET {{CarRental.Api_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
