+++ CarRental.Api/Controllers/BaseController.cs
@@ -1,5 +1,6 @@
 ﻿using CarRental.Application.Interfaces;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Infrastructure;
 
 namespace CarRental.Api.Controllers;
 
@@ -18,105 +19,103 @@ public class BaseController<TEntityCreateDTO, TEntityGetDTO>(
     /// </summary>
     [HttpPost]
     public ActionResult<int> Create([FromBody] TEntityCreateDTO entity_dto)
-    {
-        int result = -1;
-        try
+        => Log(() =>
         {
-            result = service.Create(entity_dto);
-        }
-        catch (KeyNotFoundException ex)
-        {
-            Log(404);
-            return NotFound(ex.Message);
-        }
-        Log(200);
-        return Ok(result);
-    }
+            int result = -1;
+            try
+            {
+                result = service.Create(entity_dto);
+            }
+            catch (KeyNotFoundException ex)
+            {
+                return NotFound(ex.Message);
+            }
+            return Ok(result);
+        });
 
     /// <summary>
     /// Api method for updating entity data
     /// </summary>
     [HttpPut("{id}")]
     public ActionResult Update([FromBody] TEntityCreateDTO entity_dto)
-    {
-        try
+        => Log(() =>
         {
-            service.Update(entity_dto);
-            Log(200);
-            return Ok();
-        }
-        catch (KeyNotFoundException)
-        {
-            Log(404);
-            return NotFound();
-        }
-    }
+            try
+            {
+                service.Update(entity_dto);
+                return Ok();
+            }
+            catch (KeyNotFoundException)
+            {
+                return NotFound();
+            }
+        });
 
     /// <summary>
     /// Api method for deleting entity by id
     /// </summary>
     [HttpDelete("{id}")]
     public ActionResult<bool> Delete(int id)
-    {
-        bool result = service.Delete(id);
-        if (result)
+        => Log(() =>
         {
-            Log(200);
-            return Ok();
-        }
-        Log(404);
-        return NotFound();
-    }
+            bool result = service.Delete(id);
+            if (result)
+            {
+                return Ok();
+            }
+            return NotFound();
+        });
 
     /// <summary>
     /// Api method for getting all entity instances
     /// </summary>
     [HttpGet]
     public ActionResult<List<TEntityGetDTO>> GetAll()
-    {
-        List<TEntityGetDTO> result = service.ReadAll();
-        Log(200);
-        return Ok(result);
-    }
+        => Log(() => Ok(service.ReadAll()));
+
 
     /// <summary>
     /// Api method for getting entity instance by id
     /// </summary>
     [HttpGet("{id}")]
     public ActionResult<TEntityGetDTO?> Get(int id)
+        => Log(() =>
     {
         TEntityGetDTO? result = service.Read(id);
         if (result != null)
         {
-            Log(200);
             return Ok(result);
         }
-        Log(404);
         return NotFound();
-    }
+    });
 
     /// <summary>
     /// Print data about all incoming requests
     /// </summary>
-    protected void Log(int code)
+    protected ActionResult Log(Func<ActionResult> action)
     {
         string method = HttpContext.Request.Method;
         string route = HttpContext.Request.Path;
-        DateTime timestamp = DateTime.Now;
+        string time = DateTime.Now.ToString("HH:mm:ss");
 
-        string message = $"{timestamp} {method} {route}: {code}";
+        ActionResult result;
+        int code;
 
-        if (200 <= code && code < 300)
+        try
         {
-            logger.LogInformation(message);
+            result = action();
+            code = 200;
         }
-        else if (400 <= code)
+        catch (Exception ex)
         {
-            logger.LogError(message);
+            logger.LogError(ex, $"Exception for {method} {route}");
+            result = StatusCode(500, $"{ex.Message}\n{ex.InnerException?.Message}");
+            code = 500;
         }
-        else
-        {
-            logger.LogDebug(message);
-        }
+
+        string message = $"{time} {method} {route}: {(result as IStatusCodeActionResult).StatusCode}";
+        logger.LogInformation(message);
+
+        return result;
     }
 }
