+++ CarRental.Application.Contracts/ClientCreate.cs
@@ -35,35 +35,3 @@ public class ClientCreate
 /// <summary>
 /// DTO for getting Client object
 /// </summary>
-public class ClientGet
-{
-    /// <summary>
-    /// Unique identifier for client
-    /// </summary>
-    public required int Id { get; set; }
-
-    /// <summary>
-    /// Client last name
-    /// </summary>
-    public required string LastName { get; set; }
-
-    /// <summary>
-    /// Cleint's first name
-    /// </summary>
-    public required string FirstName { get; set; }
-
-    /// <summary>
-    /// Client's patronymic
-    /// </summary>
-    public required string Patronymic { get; set; }
-
-    /// <summary>
-    /// Client's birthday date
-    /// </summary>
-    public required DateOnly BirthDate { get; set; }
-
-    /// <summary>
-    /// Client's driver license number
-    /// </summary>
-    public required string DriverLicense { get; set; }
-}
